+++ packages/genkit/test/core/action_test.dart
@@ -56,7 +56,7 @@ void main() {
     test('should start and end a span when run', () async {
       final action = Action(
         name: 'testAction',
-        actionType: 'test',
+        actionType: ActionType('test'),
         fn: (input, context) async => 'output',
       );
 
@@ -70,7 +70,7 @@ void main() {
     test('should set attributes on the span', () async {
       final action = Action(
         name: 'testAction',
-        actionType: 'test',
+        actionType: ActionType('test'),
         fn: (input, context) async => 'output',
       );
 
@@ -88,7 +88,7 @@ void main() {
     test('should run a basic action', () async {
       final action = Action(
         name: 'testAction',
-        actionType: 'test',
+        actionType: ActionType('test'),
         fn: (String? input, context) async => 'output',
       );
 
@@ -99,7 +99,7 @@ void main() {
     test('should run an action with schema', () async {
       final action = Action(
         name: 'testAction',
-        actionType: 'test',
+        actionType: ActionType('test'),
         inputSchema: TestInput.$schema,
         outputSchema: TestOutput.$schema,
         fn: (TestInput? input, context) async {
@@ -114,7 +114,7 @@ void main() {
     test('should set attributes on the span with schema', () async {
       final action = Action(
         name: 'testAction',
-        actionType: 'test',
+        actionType: ActionType('test'),
         inputSchema: TestInput.$schema,
         outputSchema: TestOutput.$schema,
         fn: (TestInput? input, context) async {
@@ -139,7 +139,7 @@ void main() {
     test('should stream an action', () async {
       final action = Action<String, String, String, void>(
         name: 'testAction',
-        actionType: 'test',
+        actionType: ActionType('test'),
         fn: (input, context) async {
           context.sendChunk('chunk1');
           context.sendChunk('chunk2');
@@ -158,7 +158,7 @@ void main() {
     test('should run an action with telemetry', () async {
       final action = Action(
         name: 'testAction',
-        actionType: 'test',
+        actionType: ActionType('test'),
         fn: (String? input, context) async => 'output',
       );
 
@@ -171,7 +171,7 @@ void main() {
     test('should run an action with provided context', () async {
       final action = Action(
         name: 'testAction',
-        actionType: 'test',
+        actionType: ActionType('test'),
         fn: (input, ctx) async {
           return ctx.context!['value'];
         },
@@ -184,14 +184,14 @@ void main() {
     test('provided context should be available in a nested action', () async {
       final innerAction = Action(
         name: 'innerAction',
-        actionType: 'test',
+        actionType: ActionType('test'),
         fn: (input, ctx) async {
           return ctx.context!['value'];
         },
       );
       final outerAction = Action(
         name: 'outerAction',
-        actionType: 'test',
+        actionType: ActionType('test'),
         fn: (input, ctx) async {
           return await innerAction(input);
         },