+++ packages/genkit_vertex_auth/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.12-rc.2
+
+ - updated internal dependencies.
+
+## 0.1.12-rc.1
+
+ - updated internal dependencies.
+
 ## 0.1.11
 
  - updated internal dependencies.